Seperate drivers from renderer and rename stdout to cli
