+++ render/driver/cli.go
@@ -1,22 +1,23 @@
-package render
+package driver
 
 import (
 	"fmt"
 	"io"
 	"os"
 
+	"github.com/sasanrose/gbench/render"
 	"github.com/sasanrose/gbench/report"
 )
 
-type stdout struct {
+type cli struct {
 	output io.Writer
 }
 
-func NewStdoutRenderer() Renderer {
-	return &stdout{os.Stdout}
+func NewCli() render.Renderer {
+	return &cli{os.Stdout}
 }
 
-func (r *stdout) Render(result *report.Result) error {
+func (r *cli) Render(result *report.Result) error {
 	tableGen := &tableGenerator{result}
 	table := tableGen.getBenchResultTable()
 	urlTables := tableGen.getUrlTables()
